pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "named-query-api"
authors = [
    {name = "Tim Holzheim", email = "tim.holzheim@rwth-aachen.de"},
]
maintainers = [
    {name = "Tim Holzheim", email = "tim.holzheim@rwth-aachen.de"},
]
readme = "README.md"
license= "Apache-2.0"

requires-python = ">=3.10"
classifiers=[
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License"
    ]

dependencies = [
    "PyYAML >= 6.0",
    "SPARQLWrapper >= 2.0.0",
    "Jinja2 >=3.1.3",
    "fastapi >= 0.109.0",
    "uvicorn >= 0.27.0",
    "dateutils"
]


dynamic = ["version", "description"]
[tool.setuptools.dynamic]
dependencies = {file = ["requirements.txt"]}
[tool.hatch.version]
path = "nqapi/__init__.py"

[project.urls]
Home = ""
Documentation = ""
Source = ""

[project.optional-dependencies]
test = [
    "httpx"
]

[tool.hatch.build.targets.wheel]
packages = [
  "nqapi",
]

[tool.isort]
profile = "black"

[tool.coverage.report]
exclude_also = [
    "if TYPE_CHECKING:"
]

[tool.tox]
legacy_tox_ini = """
    [tox]
    env_list = lint, type, py{310,311,312}
    minversion = 4.12.1

    [testenv]
    description = run the tests with pytest
    package = wheel
    wheel_build_env = .pkg
    deps =
        pytest>=6
        pytest-cov
        httpx
    commands =
        pytest --cov --cov-append {tty:--color=yes} {posargs}

    [testenv:lint]
    description = run linters
    skip_install = true
    deps =
        black==22.12
        isort
    commands =
        black {posargs:.}
        isort {posargs:.}

    [testenv:type]
    description = run type checks
    deps =
        mypy>=0.991
    commands =
        mypy {posargs:nqapi tests}
"""

[[tool.mypy.overrides]]
module = "yaml,dateutil.parser"
ignore_missing_imports = true

[project.scripts]

